rename to deal with namespace collision from cats to catz
