+++ modules/cats/src/smithy4s/catz/instances/EqInstances.scala
@@ -1,4 +1,4 @@
-package smithy4s.cats.instances
+package smithy4s.catz.instances
 
 import cats.implicits.{catsSyntaxEq, toContravariantOps}
 import cats.kernel.Eq
